requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
sqlalchemy>=2.0
psycopg2-binary>=2.9
pydantic>=2.6
python-dotenv>=1.0
openai>=1.12
openpyxl>=3.1

# --- Этап 5: веб-сервис ---
jinja2>=3.1                 # шаблоны страниц
bcrypt>=4.0                 # хеширование паролей (используется напрямую)
itsdangerous>=2.1           # подпись сессионных cookie (SessionMiddleware)
python-multipart>=0.0.9     # разбор form-data в POST-формах FastAPI
httpx>=0.27                 # клиент для тестов (FastAPI TestClient)
pytest>=8.0                 # фреймворк тестов
